Support for complete Appointment details in Get-EwsFolderItems
Get-EwsFolderItems uses the FindItem method to get the items within a folder.  This method does not return the full details for some items types, notably Appointments.  I created a switch statement within Get-EwsFolderItems prior to output that gets the folder type.  If the folder type is "IPM.Appointment", then it gets the item IDs from FindItem and uses that to call the Bind method on the Appointment type.  If this turns out to be true for other types, like Contacts or Tasks, which seems likely, then this can easily be extended to those types.